Tramento de exceção

+++ src/main/java/com/example/mercado_crud/service/UserService.java
@@ -4,6 +4,7 @@ package com.example.mercado_crud.service;
 
 import java.util.List;
 
+import com.example.mercado_crud.exceptions.RecursoNaoEncontradoException;
 import org.springframework.stereotype.Service;
 
 import com.example.mercado_crud.model.User;
@@ -34,10 +35,14 @@ public class UserService {
     }
     public User getUserByid(Long id){
         return userRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Usuario não encontrado"));
+                .orElseThrow(() -> new RecursoNaoEncontradoException("Usuario não encontrado"));
     }
 
     public void deleteUser(Long id){
+        if(!userRepository.existsById(id)){
+            throw new RecursoNaoEncontradoException("Usuario com" + id +"Não encontrado");
+        }
+
         userRepository.deleteById(id);
     }
 }
